requirements.txt
boto3~=1.26.22
botocore==1.29.62
pandas==1.4.4
numpy==1.23.4
keras==2.10.0
python-dotenv==0.21.0
tensorflow==2.10.0
tensorflow-estimator==2.10.0
scikit-learn==1.1.2
